# cmd/allternit-cloud-api/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: Allternit Platform API
  version: "1.0.0"
  summary: Agents, voice, messaging, numbers, channels and the twin layer, for developers.
  description: |
    The Allternit Platform API. Base URL `https://api.allternit.com`.

    Authenticate with a project API key: `Authorization: Bearer alt_live_…` (live
    project) or `alt_test_…` (sandbox project). Every POST accepts an
    `Idempotency-Key` header (stored per key for 24 hours). Responses carry
    `x-ratelimit-limit`, `x-ratelimit-remaining` and `x-ratelimit-reset` headers.

    Every project, sandbox included, needs a card on file (console → Billing)
    before it can do billable work. Until then billable requests answer
    `402 payment_method_required`, with the console billing page in `error.url`.

    This file is the source of truth for the `/v1` routes: a test fails if the
    router and this file disagree.
servers:
  - url: https://api.allternit.com
security:
  - bearerAuth: []
paths:
...
  /v1/numbers/{id}/consent:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      operationId: recordConsent
      summary: Record that a person agreed to be contacted
      tags: [Numbers]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [e164, source]
              properties:
                e164: { type: string }
                source: { type: string, description: "How they agreed: a form, a call, an existing customer record." }
                evidence: { type: string }
      responses:
        "201":
          description: Recorded. `opted_out` true means the person's STOP still stands.
          content:
            application/json:
              schema: { type: object, properties: { number_id: { type: string }, e164: { type: string }, recorded: { type: boolean }, opted_out: { type: boolean } } }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/messages:
    get:
      operationId: listMessages
      summary: List texts
      tags: [Messaging]
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/After" }
        - { name: number_id, in: query, schema: { type: string } }
        - { name: account_id, in: query, schema: { type: string } }
        - { name: direction, in: query, schema: { type: string, enum: [inbound, outbound] } }
      responses:
        "200":
          description: A page of messages.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: sendMessage
      summary: Send a text
      description: Refused when texting isn't active on the number, the person sent STOP, or they never texted first and have no consent recorded.
      tags: [Messaging]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [number_id, to, body], properties: { number_id: { type: string }, to: { type: string }, body: { type: string, maxLength: 1600 } } }
      responses:
        "201":
          description: The message.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        default: { $ref: "#/components/responses/Error" }
  /v1/messages/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      operationId: getMessage
      summary: Retrieve a text
      tags: [Messaging]
      responses:
        "200":
          description: The message.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A project API key, `alt_live_…` or `alt_test_…`.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size, 1 to 100 (default 20).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    After:
      name: after
      in: query
      description: Opaque cursor from `next_cursor` of the previous page.
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Up to 255 characters. Repeating a request with the same key replays the stored response for 24 hours; the same key with a different request is a 409.
      schema: { type: string, maxLength: 255 }
  responses:
    Error:
      description: Error. The HTTP status says which kind; the body is the Error object.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, code, message, param]
          properties:
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - rate_limit_error
                - conflict_error
                - billing_error
                - api_error
            code: { type: string }
            message: { type: string }
            param: { type: [string, "null"] }
            url: { type: string, description: "Where to fix it. Set on 402 payment_method_required: the console billing page for the project." }
...
    Message:
      type: object
      required: [id, object, number_id, direction, from, to, body, segments, status, created_at]
      properties:
        id: { type: string }
        object: { type: string, const: message }
        account_id: { type: [string, "null"] }
        number_id: { type: string }
        direction: { type: string, enum: [inbound, outbound] }
        from: { type: string }
        to: { type: string }
        body: { type: string }
        segments: { type: integer }
        status: { type: string, enum: [sent, simulated, received, failed] }
        error_code: { type: [string, "null"] }
        media: { type: array, items: { type: object, additionalProperties: true } }
        created_at: { type: string, format: date-time }
    MessageList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Message" } }
        has_more: { type: boolean }
        next_cursor: { type: [string, "null"] }
...